Fix: 배치 리펙토링
