--- module-batch/src/main/java/kernel/jdon/modulebatch/config/AppConfig.java
@@ -1,13 +1,0 @@
-package kernel.jdon.modulebatch.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class AppConfig {
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
