+++ module-batch/src/main/java/kernel/jdon/modulebatch/domain/jobcategory/repository/JobCategoryRepository.java
@@ -1,0 +1,10 @@
+package kernel.jdon.modulebatch.domain.jobcategory.repository;
+
+import java.util.Optional;
+
+import kernel.jdon.moduledomain.jobcategory.domain.JobCategory;
+import kernel.jdon.moduledomain.jobcategory.repository.JobCategoryDomainRepository;
+
+public interface JobCategoryRepository extends JobCategoryDomainRepository {
+    Optional<JobCategory> findByWantedCode(String wantedCode);
+}
