+++ module-batch/src/main/java/kernel/jdon/modulebatch/global/condition/SearchCondition.java
@@ -1,0 +1,7 @@
+package kernel.jdon.modulebatch.global.condition;
+
+public interface SearchCondition {
+    String getSearchValue();
+
+    String getDescription();
+}
