+++ module-batch/src/main/java/kernel/jdon/modulebatch/global/exception/BaseThrowException.java
@@ -1,0 +1,5 @@
+package kernel.jdon.modulebatch.global.exception;
+
+public interface BaseThrowException<T> {
+    T throwException();
+}
