--- module-batch/src/main/java/kernel/jdon/modulebatch/jd/reader/dto/WantedJobDetailListResponse.java
@@ -1,12 +1,0 @@
-package kernel.jdon.modulebatch.jd.reader.dto;
-
-import java.util.List;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class WantedJobDetailListResponse {
-    private final List<WantedJobDetailResponse> jobDetailList;
-}
