--- module-batch/src/main/java/kernel/jdon/modulebatch/jd/repository/JobCategoryRepository.java
@@ -1,10 +1,0 @@
-package kernel.jdon.modulebatch.jd.repository;
-
-import java.util.Optional;
-
-import kernel.jdon.moduledomain.jobcategory.domain.JobCategory;
-import kernel.jdon.moduledomain.jobcategory.repository.JobCategoryDomainRepository;
-
-public interface JobCategoryRepository extends JobCategoryDomainRepository {
-    Optional<JobCategory> findByWantedCode(String wantedCode);
-}
