--- module-batch/src/main/java/kernel/jdon/modulebatch/jd/repository/WantedJdRepository.java
@@ -1,9 +1,0 @@
-package kernel.jdon.modulebatch.jd.repository;
-
-import kernel.jdon.moduledomain.jobcategory.domain.JobCategory;
-import kernel.jdon.moduledomain.wantedjd.domain.WantedJd;
-import kernel.jdon.moduledomain.wantedjd.repository.WantedJdDomainRepository;
-
-public interface WantedJdRepository extends WantedJdDomainRepository {
-    WantedJd findByJobCategoryAndDetailId(JobCategory jobCategory, Long detailId);
-}
