--- module-batch/src/main/java/kernel/jdon/modulebatch/jd/search/SearchCondition.java
@@ -1,7 +1,0 @@
-package kernel.jdon.modulebatch.jd.search;
-
-public interface SearchCondition {
-    String getSearchValue();
-
-    String getDescription();
-}
