+++ module-batch/src/main/java/kernel/jdon/modulebatch/job/jd/reader/condition/JobSearchJobCategory.java
@@ -1,0 +1,23 @@
+package kernel.jdon.modulebatch.job.jd.reader.condition;
+
+import kernel.jdon.modulebatch.global.condition.SearchCondition;
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum JobSearchJobCategory implements SearchCondition {
+    JOB_DEVELOPER("518", "개발");
+
+    public final static String SEARCH_KEY = "job_group_id";
+    private final String searchValue;
+    private final String description;
+
+    @Override
+    public String getDescription() {
+        return this.description;
+    }
+
+    @Override
+    public String getSearchValue() {
+        return this.searchValue;
+    }
+}
