+++ module-batch/src/main/java/kernel/jdon/modulebatch/job/jd/reader/condition/JobSearchLocation.java
@@ -1,0 +1,23 @@
+package kernel.jdon.modulebatch.job.jd.reader.condition;
+
+import kernel.jdon.modulebatch.global.condition.SearchCondition;
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum JobSearchLocation implements SearchCondition {
+    LOCATIONS_ALL("all", "전체");
+
+    public final static String SEARCH_KEY = "locations";
+    private final String searchValue;
+    private final String description;
+
+    @Override
+    public String getDescription() {
+        return this.description;
+    }
+
+    @Override
+    public String getSearchValue() {
+        return this.searchValue;
+    }
+}
