+++ module-batch/src/main/java/kernel/jdon/modulebatch/job/jd/reader/dto/WantedJobListResponse.java
@@ -1,0 +1,20 @@
+package kernel.jdon.modulebatch.job.jd.reader.dto;
+
+import java.util.List;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class WantedJobListResponse {
+
+    private List<Data> data;
+
+    @Getter
+    @NoArgsConstructor(access = AccessLevel.PRIVATE)
+    public static class Data {
+        private Long id;
+    }
+}
